Got some coordinates isolated
if i check for (0,4) and (3.7), any combination of those coordinates
will work. we now know that 4 pins must be set to strong 1, and 4 set to
weak 1. then when you check if a button is pushed, you make the strong 1
a strong 0 and then see if that pulls the week 1 down to 0.